Move self-test result printing test out of main.go
